+++ qlib/rl/trainer/trainer.py
@@ -7,7 +7,7 @@ import collections
 import copy
 from contextlib import AbstractContextManager, contextmanager
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Sequence, TypeVar, cast
+from typing import Any, Dict, Iterable, List, OrderedDict, Sequence, TypeVar, cast
 
 import torch
 
@@ -151,6 +151,13 @@ class Trainer:
             "current_stage": self.current_stage,
             "metrics": self.metrics,
         }
+
+    @staticmethod
+    def get_policy_state_dict(ckpt_path: Path) -> OrderedDict:
+        state_dict = torch.load(ckpt_path, map_location="cpu")
+        if "vessel" in state_dict:
+            state_dict = state_dict["vessel"]["policy"]
+        return state_dict
 
     def load_state_dict(self, state_dict: dict) -> None:
         """Load all states into current trainer."""
